move list helpers to core, add support for task lists
