+++ packages/core/src/commands/index.ts
@@ -17,6 +17,8 @@ export * from './forEach.js'
 export * from './insertContent.js'
 export * from './insertContentAt.js'
 export * from './join.js'
+export * from './joinItemBackward.js'
+export * from './joinItemForward.js'
 export * from './keyboardShortcut.js'
 export * from './lift.js'
 export * from './liftEmptyBlock.js'
