+++ packages/core/src/commands/joinItemForward.ts
@@ -1,7 +1,19 @@
-import { RawCommands } from '@tiptap/core'
 import { joinPoint } from '@tiptap/pm/transform'
 
-export const joinListItemForward: RawCommands['splitListItem'] = () => ({
+import { RawCommands } from '../types.js'
+
+declare module '@tiptap/core' {
+  interface Commands<ReturnType> {
+    joinItemForward: {
+      /**
+       * Join two nodes Forwards.
+       */
+      joinItemForward: () => ReturnType
+    }
+  }
+}
+
+export const joinItemForward: RawCommands['joinItemForward'] = () => ({
   state,
   dispatch,
   tr,
