+++ packages/core/src/helpers/index.ts
@@ -44,6 +44,7 @@ export * from './isNodeActive.js'
 export * from './isNodeEmpty.js'
 export * from './isNodeSelection.js'
 export * from './isTextSelection.js'
+export * from './listHelpers/index.js'
 export * from './posToDOMRect.js'
 export * from './resolveFocusPosition.js'
 export * from './selectionToInsertionEnd.js'
