+++ packages/core/src/helpers/listHelpers/hasListBefore.ts
