+++ packages/core/src/helpers/listHelpers/hasListItemAfter.ts
