+++ packages/core/src/helpers/listHelpers/hasListItemBefore.ts
