+++ packages/core/src/helpers/listHelpers/listItemHasSubList.ts
@@ -1,6 +1,7 @@
-import { getNodeType } from '@tiptap/core'
 import { Node } from '@tiptap/pm/model'
 import { EditorState } from '@tiptap/pm/state'
+
+import { getNodeType } from '../getNodeType.js'
 
 export const listItemHasSubList = (typeOrName: string, state: EditorState, node?: Node) => {
   if (!node) {
