+++ packages/extension-task-item/src/task-item.ts
@@ -1,10 +1,13 @@
-import { mergeAttributes, Node, wrappingInputRule } from '@tiptap/core'
+import {
+  handleBackspace, handleDelete, KeyboardShortcutCommand, mergeAttributes, Node, wrappingInputRule,
+} from '@tiptap/core'
 import { Node as ProseMirrorNode } from '@tiptap/pm/model'
 
 export interface TaskItemOptions {
   onReadOnlyChecked?: (node: ProseMirrorNode, checked: boolean) => boolean
   nested: boolean
   HTMLAttributes: Record<string, any>
+  taskListTypeName: string
 }
 
 export const inputRegex = /^\s*(\[([( |x])?\])\s$/
@@ -16,6 +19,7 @@ export const TaskItem = Node.create<TaskItemOptions>({
     return {
       nested: false,
       HTMLAttributes: {},
+      taskListTypeName: 'taskList',
     }
   },
 
@@ -69,9 +73,15 @@ export const TaskItem = Node.create<TaskItemOptions>({
   },
 
   addKeyboardShortcuts() {
-    const shortcuts = {
+    const shortcuts: {
+      [key: string]: KeyboardShortcutCommand
+    } = {
       Enter: () => this.editor.commands.splitListItem(this.name),
       'Shift-Tab': () => this.editor.commands.liftListItem(this.name),
+      Delete: ({ editor }) => handleDelete(editor, this.name),
+      'Mod-Delete': ({ editor }) => handleDelete(editor, this.name),
+      Backspace: ({ editor }) => handleBackspace(editor, this.name, [this.options.taskListTypeName]),
+      'Mod-Backspace': ({ editor }) => handleBackspace(editor, this.name, [this.options.taskListTypeName]),
     }
 
     if (!this.options.nested) {
